requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch==2.2.0
torchvision==0.17.0
timm==0.6.13
pillow==10.0.0
python-multipart==0.0.6
